Enable segmenting for users.
When subscription is in place, users will be able to choose which tags
they want to subscribe to. Currently, the supported ones are
`lunch_meat', `dinner_meat', `lunch_veg', and `dinner_veg'.

So, they'll only receive notifications personalized for them. In the
future, likely tags include `breakfast', `pasta', and `dessert'.