Submit `let me know' form with AJAX.
The `let me know' form is now posted to `updates_subscribe'.
`update#create', which handles this, then checks whether either `ios' or
`android' is checked. If so, it then checks if the email is already in
the Updatee model. If so, it'll update the previous preferences. If not,
it'll create a new Updatee.

The home page and its relevant JS and CSS have refactored to accomodate
this. I've also simplified the CSS classes / selectors, to make it
easier to understand. (Previously, there were many similar-sounding
CSS classes / IDs.)